# Optimized Julia calculation