**Vectorization of Rust Calculations**

Vec for dynamic arrays, statically typed variables for better memory and type management, and explicit lifetime and ownership rules.

```
Old: Each loop took on average 0.1748587581 seconds to calculate.
New: Each loop took on average 0.0789778815 seconds to calculate.
```

**Vectorization of Julia Calculations**

Implemented vectorized operations to generate and evaluate all points in one go

**Efficient Timing Mechanism**

Introduced Julia's built-in @timed macro, a more efficient way to measure execution time

```
Old: Each loop took on average 0.04459999999999999 seconds to calculate.
New: Each loop took on average 0.04069297313690186 seconds to calculate.
```